refactor(lab6): current_vmm
